collector/requirements.txt
websocket_client==0.57.0
requests==2.24.0
